Fix unbound endowments
