+++ packages/snaps-execution-environments/coverage.json
@@ -1,6 +1,6 @@
 {
   "branches": 78.62,
-  "functions": 91.4,
-  "lines": 88.6,
-  "statements": 88.66
+  "functions": 91.47,
+  "lines": 88.64,
+  "statements": 88.7
 }
