+++ packages/snaps-execution-environments/src/common/endowments/index.ts
@@ -4,6 +4,7 @@ import { hasProperty } from '@metamask/utils';
 
 import { rootRealmGlobal } from '../globalObject';
 import buildCommonEndowments from './commonEndowmentFactory';
+import { bindEndowment } from './utils';
 
 type EndowmentFactoryResult = {
   /**
@@ -91,10 +92,7 @@ export function createEndowments(
         const globalValue = (rootRealmGlobal as Record<string, unknown>)[
           endowmentName
         ];
-        allEndowments[endowmentName] =
-          typeof globalValue === 'function' && !isConstructor(globalValue)
-            ? globalValue.bind(rootRealmGlobal)
-            : globalValue;
+        allEndowments[endowmentName] = bindEndowment(globalValue);
       } else {
         // If we get to this point, we've been passed an endowment that doesn't
         // exist in our current environment.
@@ -114,23 +112,4 @@ export function createEndowments(
     );
   };
   return { endowments: result.allEndowments, teardown };
-}
-
-/**
- * Checks whether the specified function is a constructor.
- *
- * @param value - Any function value.
- * @returns Whether the specified function is a constructor.
- */
-// `Function` is exactly what we want here.
-// eslint-disable-next-line @typescript-eslint/ban-types
-export function isConstructor<T extends Function>(value: T): boolean {
-  // In our current usage, the string `prototype.constructor.name` should never
-  // be empty, because you can't create a class with no name, and the
-  // `prototype.constructor.name` property is configurable but not writable.
-  // Nevertheless, that property was the empty string for `Date` in the iframe
-  // execution environment during local testing. We have no idea why, but we
-  // have to handle that case.
-  // TODO: Does the `prototype` object always have a `constructor` property?
-  return Boolean(typeof value.prototype?.constructor?.name === 'string');
 }
